Alpha 4.4 -- Still not perfect, but playable again
Fixed terribly silly error of forgetting to add the LoadSavedGame()
function to the Load Game button action. That took me a proverbial
minute to figure out (face slap)

Fixed: Player clout labeling was using the incorrect function--now uses
the new one, and added modifiers to refelct the new 5000 max clout score

Fixed: Port taxes weren't showing up properly in the notification window

Fixed: Port Taxes weren't reflected on the Port Trade GUI either

Fixed: Controls weren't being locked out in the menu system. Was causing
a lot of bad glitches where the boat would travel indefinitely or would
travel when hitting a menu button to the location behind the button.
Finally, after a looooong tortuous journey through line after line of
logic in all the linked scripts, found a way to solve this issue once
and for all. I needed a separate flag for locking controls specifically
for the menu system when it's active as well as the more generic locked
controls flag. Probably a more elegant way to handle this, but it works
quite well for now.