Pirate UI
I think at this point all the base UI is in place for the pirate game. Right now it's all full of placeholder text, but once the proper text is plugged in that should be it. It also has spaces to add the running and the negotiation code, as well as the minigame unloading.