Minor Resolution Screen Edits
The typo is fixed!

Other thing that is in the game now:

Color changing text to let the player know what resolution they are currently using.

One small issue:

The default resolution text ALWAYS starts off in green. (I'm working on getting that fixed, but feel free to look inot it and edit as you please.)