Step towards functionality for Zones
Zones now start off as inactive, then become active and then their mesh renderers are disabled so the player can not see them.

For extra measure, I added them to a new layer in the IDE called something like "Regional_Zones." No cameras in the game pick up on this new layer.

The next step for this is having the zones affect the pirate MG!

(That's probably a next week thing, though.)