Functional Heralds
Heralds work completely now, it picks a random cargo to boost and boosts it. This is reflected in the price, the "X over/under average" text, and how much you recieve when you sell. No visual indicator yet, though, and for testing it's still set to activate 100% of the time with a 100% price markup