Storm Movement
Figured out some of the issues with the storm - the upgraded ship (which was using the wrong model, actually) had its animator on doing the bobbing, which meant it couldn't move. Then, when moving it, it was going on the wrong axes. That's all fixed now, and movement shouldn't be an issue anymore